Improves post content navigation and display
Enhances the user experience for reading articles by adding scroll-to-heading functionality within the post content.

This allows users to easily navigate to specific sections of a post by clicking on links within the content.

Also, changes the logic for fetching posts for the "Latest Posts" section to retrieve articles instead of tutorials, and simplifies featured tutorials.